backend/requirements.txt
fastapi
uvicorn
crewai       
ag-ui-protocol
pandas
matplotlib
python-dotenv
